refactor: rename test files and update import paths for postman-to-bruno tests
